* Cleanup false static mock prototypes in unit tests. * Finish update of Solaris build environments. * Fix non-GCC assertion messages. * Move from Linux/BSD IP headers to PGM local defined IP headers for JSON packet dumper in network tests. * Annotate usage of _GNU_SOURCE as appropriate.

+++ trunk/openpgm/pgm/getifaddrs.c
@@ -115,7 +115,7 @@ again:
 		close (sock6);
 		return FALSE;
 	}
-	pgm_debug ("ioctl(AF_INET, SIOCGLIFCONF) = %d (%d)", lifc.lifc_len, lifc.lifc_len / sizeof(struct lifreq));
+	pgm_debug ("ioctl(AF_INET, SIOCGLIFCONF) = %d (%d)", lifc.lifc_len, (int)(lifc.lifc_len / sizeof(struct lifreq)));
 
 /* repeat everything for IPv6 */
 	lifn.lifn_family = AF_INET6;
@@ -149,7 +149,7 @@ again:
 		close (sock6);
 		return FALSE;
 	}
-	pgm_debug ("ioctl(AF_INET6, SIOCGLIFCONF) = %d (%d)", lifc6.lifc_len, lifc6.lifc_len / sizeof(struct lifreq));
+	pgm_debug ("ioctl(AF_INET6, SIOCGLIFCONF) = %d (%d)", lifc6.lifc_len, (int)(lifc6.lifc_len / sizeof(struct lifreq)));
 
 	unsigned nlifr = (lifc.lifc_len + lifc6.lifc_len) / sizeof(struct lifreq);
 	if (nlifr > if_count)
